Moved optionsHandler and getTreshold to options-handler. Removed rewire dependency
