+++ lib/output.js
@@ -2,8 +2,7 @@
 const sortOn = require('sort-on');
 const humanizeUrl = require('humanize-url');
 const prettyMs = require('pretty-ms');
-
-const THRESHOLD = 70;
+const {getThreshold} = require('./../lib/options-handler');
 
 function overview(url, strategy, scores) {
   const ret = [];
@@ -90,8 +89,6 @@ function getReporter(format) {
 }
 
 const convertToPercentum = num => num * 100;
-
-const getThreshold = threshold => (typeof threshold === 'number' && threshold && threshold < 100) ? threshold : THRESHOLD;
 
 module.exports = (parameters, response) => {
   return Promise.resolve().then(() => {
