+++ package.json
@@ -49,7 +49,6 @@
   "devDependencies": {
     "chai": "^4.2.0",
     "mocha": "^6.0.2",
-    "rewire": "^4.0.1",
     "xo": "^0.25.3"
   },
   "xo": {
