+++ test/index.test.js
@@ -1,9 +1,5 @@
 /* eslint-env mocha */
 const {expect} = require('chai');
-const rewire = require('rewire');
-
-const psiWired = rewire('..');
-const handleOptions = psiWired.__get__('handleOpts');
 
 const psi = require('..');
 
@@ -14,26 +10,5 @@ describe('Index file', () => {
       .catch(error => {
         expect(error.message).to.eql('URL required');
       });
-  });
-  describe('handleOpts method', async () => {
-    const defaultConfig = {
-      nokey: true,
-      url: 'http://addyosmani.com'
-    };
-
-    it('should return strategy "mobile" if no strategy is passed', () => {
-      const expectedOutput = {
-        ...defaultConfig,
-        strategy: 'mobile'
-      };
-      expect(handleOptions('http://addyosmani.com')).to.eql(expectedOutput);
-    });
-    it('should respect passed strategy options', () => {
-      const expectedOutput = {
-        ...defaultConfig,
-        strategy: 'desktop'
-      };
-      expect(handleOptions('http://addyosmani.com', {strategy: 'desktop'})).to.eql(expectedOutput);
-    });
   });
 });
